+++ bench/random_read/main.c
@@ -44,7 +44,7 @@ sendbatch(struct randomread_state * C)
 
 	while (C->Nip < 4096) {
 		/* Generate a random key. */
-		N = random() % C->Nmax;
+		N = (size_t)random() % C->Nmax;
 		X = N >> 16;
 		Y = N - (X << 16);
 		mkkey(X, Y, C->key->buf);
